Added platforms and floor and their physics implementations.
Platforms are spawned in RacerArea.spawnPlatform(length, lane, xCord). This allows them to be given a length (at the moment this just means to spawn multiple platforms next to each other) and spawn at a vector (xCord, lane). The lane should be one of the variables set as LANE_1, LANE_2, ... to keep the heights regular. Some example platforms are spawned.

The floor is spawned using RacerArea.spawnFloor(), which at the moment simply spawns enough floor to cover the ground.

ObstacleFactory.createPlatform() and ObstacleFactory.createFloor() add the components:
 - TextureRenderComponent,
 - PhysicsComponent as a static body, and
 - ColliderComponent with static body type
 to the entities.