HOW TO ACCESS OBSTACLE GENERATION MODE: Go into MainGameScreen Set boolean variable "isObstacle" to true Then you should be in the mode Otherwise you can just play the old version.
MAIN COMMIT: Added in Obstacle Generation and a Debug Screen.

Added RAGNAROK_MAIN enum value in Terrain Factory.

Added ObstacleTypes to determine which obstacles to spawn.

Added ObstacleArea to model endless runner, randomly generated obstacles.

Added NPCConfigs for rocks and spikes.

Added GeneratorComponent, an entity component that is passed the ObstacleArea instance and spawns enemies in the world randomly. The logic for the generation is contained in here, if you like to read that sorta stuff.

Added GeneratorFactory that makes entities conforming to ObstacleArea standards.

Added more Debug Information to GameAreaDisplay. Can be turned off by setting the "isDebug" boolean variable to false;

Added a song.