Commit 5/9/21 11:30
Have moved a bunch of methods from RacerArea into new class RagnarokArea, just for the sake of resetting my brain. Now I'm going to work on adding all the entity creation methods into RagnarokArea so they can be interfaced to from the Terminal through the AreaManager class (which is a subclass of RagnarokArea, so it can call protected functions).

todo:
 [x] move entity
 [ ] interface to terminal
 [ ] know of area stuff
 [ ] implement -load into terminal
 [ ]