Implemented the Wall of Death disposing of entities.
The component TouchDisposeComponent has been implemented, which has been added to the Wall of Death. After the wall of death passes an object, the object will be disposed. After it passes an NPC, the NPC will be disabled. This is because if the animation component of an NPC is disposed it will dispose the atlas file, making all other entities using it turn black.

The entities are disposed of out of step with the physics engine using the entity service. If they were disposed of inside the physics step the game crashes.