AreaManger has column-wise sorted array ragEdit can be found in root directory
Further notes:
Check the AreaManager for a String[][] of "-place" commands that are indexed first by [x] (column-wise).

Level Editor doesn't have new load but editing and saving works a-okay, you just have to call the actual name of the file in the program, not even filepath. i.e. if u save as "configs/wickedshit.rag", you just call "-load wickedshit" in the game