Added ragEditv1.3.jar and the ragEdit Tutorial
Updated ragEdit to latest version (1.3). It may still break in the coming weeks as features are asymetrically added across the game... but please keep on the pulse of it. If it's not working as expected, let me know ASAP!, please document your errors. If you wish to be added to the source, let me know.