Part way through refactoring the SoundService.
Todo is scope out the on hit sound effect (I got no clue where it is) and also to mess with the giant distance stuff. Otherwise, it's not so bhed.

Have to test the single instance of music to see if tracks can change w/out messing things up. Also mess w/ multi tracked stuff.